Added config inspection.
Improved default inspection by answering computed key and value in
shell format. Example:

    <key>=<value>

Additionally, ensures the entire environment is not output when
constructing a new configuration.

Rubocop and Reek violations will be addressed in the future since they
can't be delt with now without forcing a major release of the gem.