Updated friendly path as home path.
The original object name was slightly confusing as it was *friendly* in
the sense of being friendly for development purposes. Yet, it didn't
fully represent what this path object truly represents which is a
preference for the local versus global home path. This change makes
this distinction more clear and mimics changes in the XDG gem.