Fixed current common path to only detect files
Necessary to ensure directory paths that exist but are empty are not
answered back. This method is purely file focused so detecting an
existing directory is a false positive and can cause downstream objects
to not load the files they need.

An example of this can be found in the `Config` object where, upon
attempting to load settings, an existing directory would cause an empty
hash to be loaded instead of finding the actual configuration.