Updated common path to use active instead of current terminology
The change in terminology is being introduced in order for the method to have more descriptive meaning since the *active* path is the path to an actual file which exists versus a path which may or may not exist. The previous use of *current* wasn't as clear in this regard.