Updated context to be a data object
Necessary to simplify the logic since `Data` is a better implementation instead of a using a `Struct`. Additionally, the ability to supply a class or instance of an XDG object is now supported for improved customization. This will soon be used by the `Runcom::Data` and `Runcom::State` objects.

Milestone: minor